Simplifed tests structure and reduced duplicates
