+++ tests/smoke/test_canvas.py
